Add code from osu! for slider path calcs (currently unused), refactoring
- Note that some code was adapted, but 90% is straight from ppy/osu or ppy/osu-framework
- double -> float for positions since they limited it

+++ OsuDiffCalc/FileProcessor/BeatmapObjects/HitObject.cs
@@ -6,15 +6,15 @@
 	/// An interactable object in a beatmap with a time and position (eg. everything but a break section)
 	/// </summary>
 	abstract class HitObject : BeatmapObject {
-		public HitObject(double x, double y, double startTime, double endTime) : base(startTime, endTime) {
+		public HitObject(float x, float y, double startTime, double endTime) : base(startTime, endTime) {
 			X = x;
 			Y = y;
 		}
 
 		/// <summary> X initial position of the HitObject in osupixels </summary>
-		public double X { get; protected init; }
+		public float X { get; protected init; }
 		/// <summary> Y initial position of the HitObject in osupixels </summary>
-		public double Y { get; protected init; }
+		public float Y { get; protected init; }
 
 		public override void PrintDebug(string prepend = "", string append = "") {
 			Console.Write(prepend);
