+++ OsuDiffCalc/FileProcessor/BeatmapObjects/Spinner.cs
@@ -1,7 +1,7 @@
 ﻿
 namespace OsuDiffCalc.FileProcessor.BeatmapObjects {
 	class Spinner : HitObject {
-		public Spinner(double x, double y, double startTime, double endTime) : base(x, y, startTime, endTime) {
+		public Spinner(float x, float y, double startTime, double endTime) : base(x, y, startTime, endTime) {
 		}
 	}
 }
